+++ seleniumtid/pageelements/page_element.py
@@ -1,14 +1,16 @@
 # -*- coding: utf-8 -*-
-'''
-(c) Copyright 2014 Telefonica, I+D. Printed in Spain (Europe). All Rights
+
+u"""
+(c) Copyright 2014 Telefónica, I+D. Printed in Spain (Europe). All Rights
 Reserved.
 
-The copyright to the software program(s) is property of Telefonica I+D.
+The copyright to the software program(s) is property of Telefónica I+D.
 The program(s) may be used and or copied only with the express written
-consent of Telefonica I+D or in accordance with the terms and conditions
+consent of Telefónica I+D or in accordance with the terms and conditions
 stipulated in the agreement/contract under which the program(s) have
 been supplied.
-'''
+"""
+
 from seleniumtid import selenium_driver
 
 
@@ -19,10 +21,12 @@ class PageElement(object):
 
     @property
     def driver(self):
-        '''
-        This method allows to autocomplete self.driver in IDEs
-        :rtype selenium.webdriver.remote.webdriver.WebDriver
-        '''
+        """Get the Selenium driver
+         This method allows to autocomplete self.driver in IDEs
+
+        :returns: Selenium driver
+        :rtype: selenium.webdriver.remote.webdriver.WebDriver
+        """
         return self._driver
 
     def set_driver(self, driver):