+++ seleniumtid/pageelements/select_page_element.py
@@ -1,16 +1,19 @@
 # -*- coding: utf-8 -*-
-'''
-(c) Copyright 2014 Telefonica, I+D. Printed in Spain (Europe). All Rights
+
+u"""
+(c) Copyright 2014 Telefónica, I+D. Printed in Spain (Europe). All Rights
 Reserved.
 
-The copyright to the software program(s) is property of Telefonica I+D.
+The copyright to the software program(s) is property of Telefónica I+D.
 The program(s) may be used and or copied only with the express written
-consent of Telefonica I+D or in accordance with the terms and conditions
+consent of Telefónica I+D or in accordance with the terms and conditions
 stipulated in the agreement/contract under which the program(s) have
 been supplied.
-'''
-from seleniumtid.pageelements.page_element import PageElement
+"""
+
 from selenium.webdriver.support.ui import Select as SeleniumSelect
+
+from seleniumtid.pageelements.page_element import PageElement
 
 
 class Select(PageElement):