+++ seleniumtid/test/test_selenium_wrapper.py
@@ -1,25 +1,29 @@
 # -*- coding: utf-8 -*-
-'''
-(c) Copyright 2015 Telefonica, I+D. Printed in Spain (Europe). All Rights
+
+u"""
+(c) Copyright 2015 Telefónica, I+D. Printed in Spain (Europe). All Rights
 Reserved.
 
-The copyright to the software program(s) is property of Telefonica I+D.
+The copyright to the software program(s) is property of Telefónica I+D.
 The program(s) may be used and or copied only with the express written
-consent of Telefonica I+D or in accordance with the terms and conditions
+consent of Telefónica I+D or in accordance with the terms and conditions
 stipulated in the agreement/contract under which the program(s) have
 been supplied.
-'''
+"""
+
 import unittest
-from seleniumtid.config_parser import ExtendedConfigParser
-from seleniumtid.selenium_wrapper import SeleniumWrapper
-from ddt import ddt, data, unpack
 import logging
 import os
+
+from seleniumtid.selenium_wrapper import SeleniumWrapper
+from ddt import ddt, data, unpack
 import mock
 
 
 class SeleniumWrapperCommon(unittest.TestCase):
-    conf_folder = 'seleniumtid/test/conf/'
+    @classmethod
+    def setUpClass(cls):
+        cls.root_path = os.path.dirname(os.path.realpath(__file__))
 
     def tearDown(self):
         # Remove environment properties
@@ -35,19 +39,19 @@ class SeleniumWrapperCommon(unittest.TestCase):
 
 class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
     def setUp(self):
-        os.environ["Files_logging"] = self.conf_folder + 'logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'logging.conf')
         self.wrapper = SeleniumWrapper()
         self.wrapper.configure_logger()
 
     def test_configure_properties(self):
-        os.environ["Files_properties"] = self.conf_folder + 'properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'properties.cfg')
         self.wrapper.configure_properties()
         self.assertEquals('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
         self.assertEquals('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
         self.assertEquals(None, self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_android(self):
-        os.environ["Files_properties"] = self.conf_folder + 'android-properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'android-properties.cfg')
         self.wrapper.configure_properties()
         self.assertEquals('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
         self.assertEquals(None, self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
@@ -55,8 +59,8 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
                           self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_two_files(self):
-        os.environ["Files_properties"] = (self.conf_folder + 'properties.cfg;' + self.conf_folder +
-                                          'android-properties.cfg')
+        os.environ["Files_properties"] = (os.path.join(self.root_path, 'conf', 'properties.cfg') + ';' +
+                                          os.path.join(self.root_path, 'conf', 'android-properties.cfg'))
         self.wrapper.configure_properties()
         self.assertEquals('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
         self.assertEquals('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
@@ -64,8 +68,8 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
                           self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_two_files_android_first(self):
-        os.environ["Files_properties"] = (self.conf_folder + 'android-properties.cfg;' + self.conf_folder +
-                                          'properties.cfg')
+        os.environ["Files_properties"] = (os.path.join(self.root_path, 'conf', 'android-properties.cfg') + ';' +
+                                          os.path.join(self.root_path, 'conf', 'properties.cfg'))
         self.wrapper.configure_properties()
         self.assertEquals('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
         self.assertEquals('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
@@ -73,7 +77,7 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
                           self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_system_property(self):
-        os.environ["Files_properties"] = self.conf_folder + 'properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'properties.cfg')
         os.environ["Browser_browser"] = 'opera'
         self.wrapper.configure_properties()
         self.assertEquals('opera', self.wrapper.config.get('Browser', 'browser'))
@@ -83,12 +87,12 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
         self.assertRaisesRegexp(Exception, 'Properties config file not found', self.wrapper.configure_properties)
 
     def test_configure_properties_file_not_found(self):
-        os.environ["Files_properties"] = self.conf_folder + 'notfound-properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'notfound-properties.cfg')
         self.assertRaisesRegexp(Exception, 'Properties config file not found', self.wrapper.configure_properties)
 
     def test_configure_properties_no_changes(self):
         # Configure properties
-        os.environ["Files_properties"] = self.conf_folder + 'properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'properties.cfg')
         self.wrapper.configure_properties()
         self.assertEquals('firefox', self.wrapper.config.get('Browser', 'browser'))
 
@@ -105,7 +109,7 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
 
     def test_configure_properties_change_configuration_file(self):
         # Configure properties
-        os.environ["Files_properties"] = self.conf_folder + 'properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'properties.cfg')
         self.wrapper.configure_properties()
         self.assertEquals('firefox', self.wrapper.config.get('Browser', 'browser'))
 
@@ -115,7 +119,7 @@ class SeleniumWrapperPropertiesTests(SeleniumWrapperCommon):
         self.assertEquals(new_browser, self.wrapper.config.get('Browser', 'browser'))
 
         # Change file and try to configure again
-        os.environ["Files_properties"] = self.conf_folder + 'android-properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'android-properties.cfg')
         self.wrapper.configure_properties()
 
         # Check that configuration has been initialized
@@ -127,7 +131,7 @@ class SeleniumWrapperLoggerTests(SeleniumWrapperCommon):
         self.wrapper = SeleniumWrapper()
 
     def test_configure_logger(self):
-        os.environ["Files_logging"] = self.conf_folder + 'logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'logging.conf')
         self.wrapper.configure_logger()
         self.assertEquals('DEBUG', logging.getLevelName(self.wrapper.logger.getEffectiveLevel()))
 
@@ -136,13 +140,13 @@ class SeleniumWrapperLoggerTests(SeleniumWrapperCommon):
         self.wrapper.logger.info('No error, default logging configuration')
 
     def test_configure_logger_file_not_found(self):
-        os.environ["Files_logging"] = self.conf_folder + 'notfound-logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'notfound-logging.conf')
         self.wrapper.configure_logger()
         self.wrapper.logger.info('No error, default logging configuration')
 
     def test_configure_logger_no_changes(self):
         # Configure logger
-        os.environ["Files_logging"] = self.conf_folder + 'logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'logging.conf')
         self.wrapper.configure_logger()
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
         self.assertEquals('DEBUG', logging.getLevelName(logger.level))
@@ -160,7 +164,7 @@ class SeleniumWrapperLoggerTests(SeleniumWrapperCommon):
 
     def test_configure_logger_change_configuration_file(self):
         # Configure logger
-        os.environ["Files_logging"] = self.conf_folder + 'logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'logging.conf')
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
         self.assertEquals('DEBUG', logging.getLevelName(logger.level))
 
@@ -170,7 +174,7 @@ class SeleniumWrapperLoggerTests(SeleniumWrapperCommon):
         self.assertEquals(new_level, logging.getLevelName(logger.level))
 
         # Change file and try to configure again
-        os.environ["Files_logging"] = self.conf_folder + 'warn-logging.conf'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'warn-logging.conf')
         self.wrapper.configure_logger()
 
         # Check that configuration has been initialized
@@ -210,8 +214,8 @@ maximizable_browsers = (
 @ddt
 class SeleniumWrapperTests(SeleniumWrapperCommon):
     def setUp(self):
-        os.environ["Files_logging"] = self.conf_folder + 'logging.conf'
-        os.environ["Files_properties"] = self.conf_folder + 'properties.cfg'
+        os.environ["Files_logging"] = os.path.join(self.root_path, 'conf', 'logging.conf')
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'properties.cfg')
         self.wrapper = SeleniumWrapper()
         self.wrapper.configure()
 
@@ -250,7 +254,7 @@ class SeleniumWrapperTests(SeleniumWrapperCommon):
         self.wrapper.screenshots_number += 1
 
         # Change browser and try to configure again
-        os.environ["Files_properties"] = self.conf_folder + 'android-properties.cfg'
+        os.environ["Files_properties"] = os.path.join(self.root_path, 'conf', 'android-properties.cfg')
         self.wrapper.configure()
 
         # Check that configuration has been initialized