+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 seleniumtid changelog
 =====================
 
+v0.9.0
+------
+
+*Release date: 2015-05-12*
+
+  * Output path (screenshots, videos, visualtests) can be specified with a system property: *Files_output_path*
+  * Update app_strings in Appium tests only if the driver has changed
+  * Move visual properties from [Server] section to [VisualTests] section
+  * With a visual assertion error, the test can fail or give an error message and continue
+    New config property 'fail' in [VisualTests] section to fail the test when there is a visual error
+  * Create a html report with the visual tests results
+    New config property 'complete_report' in [VisualTests] section to include also correct visual assertions in report
+  * Configure multiple baseline name for different browsers, languages and versions
+    New config property 'baseline_name' in [VisualTests] section to configure the name of the baseline folder
+    Allow {browser}, {language} and {platformVersion} variables, i.e. baseline_name = {browser}-{language}. The default baseline_name is {browser}.
+  * Add assertFullScreenshot method in SeleniumTestCase
+
 v0.8.6
 ------